update argument types for latest api
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@29811 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSObjCRuntime.m
@@ -134,7 +134,8 @@ NSStringFromClass(Class aClass)
  * The function returns a pointer to the type information C string.
  */
 const char *
-NSGetSizeAndAlignment(const char *typePtr, unsigned *sizep, unsigned *alignp)
+NSGetSizeAndAlignment(const char *typePtr,
+  NSUInteger *sizep, NSUInteger *alignp)
 {
   typePtr = objc_skip_offset (typePtr);
   typePtr = objc_skip_type_qualifiers (typePtr);
